Styling is good enough
Yes, I know the progress view can be initialized with 11/10 at the end of the game and that's a runtime error. I tried to clamp it directly in the initializer and that didn't work, but I've also been sitting on this for two days and I want to move on to the next project.